Split the legacy StatisticV2 pipeline out of dataProcessing
- processData/processForChart/createProcessor/filterByDateRange and the
  class re-exports move to legacy-transform.js with a warning about the
  positional index-alignment bug in TimeSeriesV2.transform(); the current
  chart routes bypass this pipeline entirely via the series-rows core
- dataProcessing.js keeps only the display aggregation and re-exports the
  legacy module, so v2/index.js and studio/lens-on-au-grid are untouched
